KTOR-6851 Parsing Content-Length to Long instead of Int to prevent Nu… (#4008)
* KTOR-6851 Parse Content-Length to Long instead of Int to prevent NumberFormatException for lengths more than Int.MAX_VALUE

(cherry picked from commit 5513dcfa14dc94491f48ca7e6b2eebc3f33c8821)

+++ ktor-server/ktor-server-cio/jvmAndNix/src/io/ktor/server/cio/CIOApplicationEngine.kt
@@ -151,7 +151,7 @@ public class CIOApplicationEngine(
     }
 
     private fun hasBody(request: CIOApplicationRequest): Boolean {
-        val contentLength = request.headers[HttpHeaders.ContentLength]?.toInt()
+        val contentLength = request.headers[HttpHeaders.ContentLength]?.toLong()
         val transferEncoding = request.headers[HttpHeaders.TransferEncoding]
         return transferEncoding != null || (contentLength != null && contentLength > 0)
     }
